backend/requirements.txt
fastapi             ~=0.110
pymongo             ~=4.13
uvicorn             ~=0.28
pydantic_settings
pydantic[email]
passlib
pinecone[asyncio]==6.0.2
python-jose[cryptography]>=3.3.0
ollama>=0.1.8
bcrypt